+++ src/commonMain/kotlin/core/utility/KotlinResourceHelper.kt
@@ -1,5 +1,5 @@
 package core.utility
 
-expect class KotlinResourceHelper : ResourceHelper {
+expect class KotlinResourceHelper() : ResourceHelper {
     override fun getResourceFiles(path: String, recursive: Boolean): List<String>
 }
