+++ src/jsMain/kotlin/core/utility/SystemUtil.kt
@@ -1,8 +1,10 @@
 package core.utility
 
+import core.events.EventListener
 import io.ktor.client.*
 import io.ktor.client.plugins.contentnegotiation.*
 import io.ktor.serialization.kotlinx.json.*
+import kotlin.reflect.KClass
 
 actual fun currentTime(): Long {
     throw NotImplementedError()
@@ -24,6 +26,10 @@ actual fun buildWebClient() : HttpClient {
     TODO("Not yet implemented")
 }
 
+actual fun getListenedForClass(listener: EventListener<*>): KClass<*> {
+    TODO("Not yet implemented")
+}
+
 actual object Math {
     actual fun random(): Double {
         TODO("Not yet implemented")
@@ -41,7 +47,7 @@ actual object Math {
     }
 }
 
-object Integer {
+actual object Integer {
     actual fun parseInt(value: String?): Int {
         throw NotImplementedError()
     }
